feat:카테고리, 태그, 태그별장소 엔티티/VO 분리
