+++ src/main/java/com/ohgiraffers/adregamdi/place/command/domain/aggregate/vo/CoordinateVO.java
@@ -20,13 +20,13 @@ public class CoordinateVO {
 
     public CoordinateVO(double latitude, double longitude) {
 
-        if (latitude >= 33 && latitude <= 34){
+        if (latitude >= 32 && latitude <= 35){
             this.latitude = latitude;
         } else {
             this.latitude = 0;
         }
 
-        if (longitude >= 125 && longitude <= 127){
+        if (longitude >= 124 && longitude <= 128){
             this.longitude = longitude;
         } else {
             this.longitude = 0;
