+++ src/main/java/com/ohgiraffers/adregamdi/place/command/domain/service/PlaceAPIService.java
@@ -1,8 +1,8 @@
 package com.ohgiraffers.adregamdi.place.command.domain.service;
 
+import com.ohgiraffers.adregamdi.category.command.domain.aggregate.entity.Category;
 import com.ohgiraffers.adregamdi.place.command.domain.aggregate.entity.*;
-
-import java.util.List;
+import com.ohgiraffers.adregamdi.tag.command.domain.aggregate.entity.Tag;
 
 public interface PlaceAPIService {
