+++ src/main/java/com/ohgiraffers/adregamdi/place/query/application/service/DataQueryService.java
@@ -1,11 +1,11 @@
 package com.ohgiraffers.adregamdi.place.query.application.service;
 
+import com.ohgiraffers.adregamdi.category.command.domain.aggregate.entity.Category;
 import com.ohgiraffers.adregamdi.place.command.domain.aggregate.entity.*;
 import com.ohgiraffers.adregamdi.place.query.infra.repository.*;
+import com.ohgiraffers.adregamdi.tag.command.domain.aggregate.entity.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 public class DataQueryService {
