+++ src/main/java/com/ohgiraffers/adregamdi/place/query/infra/repository/CategoryQueryRepository.java
@@ -1,6 +1,6 @@
 package com.ohgiraffers.adregamdi.place.query.infra.repository;
 
-import com.ohgiraffers.adregamdi.place.command.domain.aggregate.entity.Category;
+import com.ohgiraffers.adregamdi.category.command.domain.aggregate.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryQueryRepository extends JpaRepository<Category, Long> {
