+++ src/main/java/com/ohgiraffers/adregamdi/place/query/infra/repository/TagQueryRepository.java
@@ -1,6 +1,6 @@
 package com.ohgiraffers.adregamdi.place.query.infra.repository;
 
-import com.ohgiraffers.adregamdi.place.command.domain.aggregate.entity.Tag;
+import com.ohgiraffers.adregamdi.tag.command.domain.aggregate.entity.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TagQueryRepository extends JpaRepository<Tag, Long> {
