+++ src/main/java/com/ohgiraffers/adregamdi/tagplace/command/domain/aggregate/vo/PlaceVO.java
@@ -1,4 +1,4 @@
-package com.ohgiraffers.adregamdi.place.command.domain.aggregate.vo;
+package com.ohgiraffers.adregamdi.tagplace.command.domain.aggregate.vo;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -11,8 +11,9 @@ import javax.persistence.Embeddable;
 @NoArgsConstructor
 @AllArgsConstructor
 @Embeddable
-public class TagVO {
+public class PlaceVO {
 
-    @Column(nullable = false, name = "tag_no")
-    private Long tagNo;
+    @Column(nullable = false, name = "place_no")
+    private Long placeNo;
+
 }
